Working through upgrades (ruby 2.7.1, sinatra 2.1, AR 5.2)
My eventual goal is to get this app upgraded to ruby 3.0, but it's been
a while since I tried to actively maintain all the gem dependencies. To
tackle this, I'm going to start by upgrading to ruby 2.7.1 since it
should be the "noisy" version that alerts for a bunch of deprecations
that are likely to break things in 3.0 (primarily the changes around
mixing hash and keywork args). Sadly, it looks like these upgrades are
going to be a bit gnarly, so saving a checkpoint here since I think
things are working. Moving to sinatra-activerecord newer than 2.0.17
seems to introduce some config loading incompatibilities in rake tasks,
at least with AR pre-6.x. AR 5.2 is incompatible with ruby 3 due to the
keywork arg changes, but AR 6.x seems to have other config
loading/definition issues I'm going to have to debug and work through.

Switch from travis-ci to circleci