Fighting with db configuration in AR 6.0
Think I've managed to hack around the various challenges introduced by
newer versions of ActiveRecord that make it more difficult to integrate
with outside of a Rails project. Simplified database.yml to have a
single db config and try to automatically match the current env.

Upgrades to latest ActiveRecord 6.0 and sinatra-activerecord, which
should hopefully unblock upgrading to Ruby 3.0.